+++ features-rlgym-rocket-league/src/features_rlgym/rocket_league/features/feature_self.py
@@ -48,9 +48,7 @@ class FeatureSelf(Feature[Hashable, np.ndarray, Any, GameState, int]):
                 _physics.up,
             ]
 
-            _new_obs[_agent] = np.concatenate(
-                (_existing_obs, *_added_obs), dtype=_existing_obs.dtype
-            )
+            _new_obs[_agent] = np.concatenate((_existing_obs, *_added_obs))
 
         return _new_obs
 
@@ -70,9 +68,11 @@ def add_self_feature(
     :type config: FeatureConfig
     :param position_normalization: Position normalization coefficient, defaults to 1/2300
     :type position_normalization: float | np.ndarray, optional
-    :param linear_velocity_normalization: Linear velocity normalization coefficient, defaults to 1/2300
+    :param linear_velocity_normalization: Linear velocity normalization
+        coefficient, defaults to 1/2300
     :type linear_velocity_normalization: float | np.ndarray, optional
-    :param angular_velocity_normalization: Angular velocity normalization coefficient, defaults to 1/math.pi
+    :param angular_velocity_normalization: Angular velocity normalization
+        coefficient, defaults to 1/math.pi
     :type angular_velocity_normalization: float | np.ndarray, optional
     """
     config.obs_builder.add_feature(
